mechanization of progress/preservation proofs for System F and also one proof for 23.6.3 (2) from TAPL (which needs preservation)
Preservation is not completely finished yet because it needs the context
morphism lemma (but that should hold).
Maybe when my autosubst implementation is fully featured I can prove it.
As of yet it should be possible by copying Kathrin's proof for the
Poplmark challenge and changing the parts where she uses the asimpl tactic.